feat #27 : 공용 컴포넌트 분리
