+++ src/components/detailProfile/DetailCard.module.scss
